Make 1500 MTU work over DSL
the modem seems to be able to both send and receive 1508 ethernet
packets, even though it's network interface is set to 1500. I guess
that's due to the way they forward packets between dsl and ethernet.

So all we have to do is increase the MTU on our side as well and
configure ppp accordingly. I tested this with non fragmentable pings:
ping -M do -s 1472 1.1.1.1

This works just fine now.