add wg-gen-client-config
This can generate a config for any client, which can also be piped into
qrencode and scanned by a smartphone. The PrivateKey is all zeroes as
it's intended to be generated on the target device, so I can then copy
it's public key into the config on the router. This makes sure the
private key never leaves the target device.